Server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
joblib==1.3.2
imbalanced-learn==0.11.0
xgboost==1.7.6
lightgbm==4.1.0
tensorflow==2.13.0
sqlalchemy==2.0.21
python-dotenv==1.0.0
